+++ ZooSim.App/Monkey.cs
@@ -2,11 +2,11 @@
 {
     public class Monkey : Animal
     {
-        private const int MortalityPercentage = 30;
+        private const float MortalityPercentage = 30;
         public Monkey() : base(MortalityPercentage)
         {
             Type = "Monkey";
-            Health = 100;
+            Health = 100f;
         }
 
     }
